Move and rename service worker files
