+++ site/source/Provider.tsx
@@ -18,9 +18,9 @@ import { Link } from '@/design-system/typography/link'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
 import { useIframeResizer } from '@/hooks/useIframeResizer'
 
-import { ServiceWorker } from './ServiceWorker'
 import { TrackingContext } from './components/ATInternetTracking'
 import { createTracker } from './components/ATInternetTracking/Tracker'
+import { ServiceWorker } from './components/ServiceWorker'
 import { DarkModeProvider } from './contexts/DarkModeContext'
 import { Message } from './design-system'
 import * as safeLocalStorage from './storage/safeLocalStorage'
