+++ site/source/components/ServiceWorker.tsx
@@ -5,9 +5,9 @@ import { useRegisterSW } from 'virtual:pwa-register/react'
 
 import { getItem, removeItem, setItem } from '@/storage/safeLocalStorage'
 
-import { Message } from './design-system'
-import { Button, CloseButton } from './design-system/buttons'
-import { Body } from './design-system/typography/paragraphs'
+import { Message } from '../design-system'
+import { Button, CloseButton } from '../design-system/buttons'
+import { Body } from '../design-system/typography/paragraphs'
 
 const PromptContainer = styled.div`
 	position: fixed;
